% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRequiredProbes.R
\name{getRequiredProbes}
\alias{getRequiredProbes}
\title{See the list of probes needed for each Dunedin Pace of Aging Methylation Score Model}
\usage{
getRequiredProbes()
}
\value{
A list of string vectors.  Each element in the list is a different model, and the character vector contains the IDs needed for that particular model.
}
\description{
\code{getRequiredProbes} returns a list containing the IDs of the probes needed for each Dunedin Pace of Aging Methylation Score Model.
}
\examples{
getRequiredProbes()

}
